UPSTREAM: 52073: Fix cross-build
:100644 100644 9f4753b222... c5de9dad14... M	pkg/kubelet/cadvisor/BUILD
:100644 100644 7dbbde83ff... be42d39e36... M	pkg/kubelet/cadvisor/cadvisor_unsupported.go
:100644 100644 0033f92314... 404dd05c75... M	pkg/kubelet/cadvisor/cadvisor_windows.go
:000000 100644 0000000000... a9e88b5bf8... A	pkg/kubelet/cadvisor/helpers_linux.go
:000000 100644 0000000000... af90f150c1... A	pkg/kubelet/cadvisor/helpers_unsupported.go
:100644 100644 42248f6695... bd30c66530... M	pkg/kubelet/cadvisor/util.go
