+++ pkg/kubelet/cadvisor/cadvisor_unsupported.go
@@ -80,3 +80,7 @@ func (cu *cadvisorUnsupported) WatchEvents(request *events.Request) (*events.Eve
 func (cu *cadvisorUnsupported) HasDedicatedImageFs() (bool, error) {
 	return false, unsupportedErr
 }
+
+func (c *cadvisorUnsupported) GetFsInfoByFsUUID(uuid string) (cadvisorapiv2.FsInfo, error) {
+	return cadvisorapiv2.FsInfo{}, nil
+}
