+++ pkg/kubelet/cadvisor/util.go
@@ -17,43 +17,11 @@ limitations under the License.
 package cadvisor
 
 import (
-	"fmt"
-
-	cadvisorfs "github.com/google/cadvisor/fs"
 	cadvisorapi "github.com/google/cadvisor/info/v1"
 	cadvisorapi2 "github.com/google/cadvisor/info/v2"
 	"k8s.io/apimachinery/pkg/api/resource"
 	"k8s.io/kubernetes/pkg/api/v1"
 )
-
-// imageFsInfoProvider knows how to translate the configured runtime
-// to its file system label for images.
-type imageFsInfoProvider struct {
-	runtime         string
-	runtimeEndpoint string
-}
-
-// ImageFsInfoLabel returns the image fs label for the configured runtime.
-// For remote runtimes, it handles additional runtimes natively understood by cAdvisor.
-func (i *imageFsInfoProvider) ImageFsInfoLabel() (string, error) {
-	switch i.runtime {
-	case "docker":
-		return cadvisorfs.LabelDockerImages, nil
-	case "rkt":
-		return cadvisorfs.LabelRktImages, nil
-	case "remote":
-		// TODO: pending rebase including https://github.com/google/cadvisor/pull/1741
-		if i.runtimeEndpoint == "/var/run/crio.sock" {
-			return "crio-images", nil
-		}
-	}
-	return "", fmt.Errorf("no imagefs label for configured runtime")
-}
-
-// NewImageFsInfoProvider returns a provider for the specified runtime configuration.
-func NewImageFsInfoProvider(runtime, runtimeEndpoint string) ImageFsInfoProvider {
-	return &imageFsInfoProvider{runtime: runtime, runtimeEndpoint: runtimeEndpoint}
-}
 
 func CapacityFromMachineInfo(info *cadvisorapi.MachineInfo) v1.ResourceList {
 	c := v1.ResourceList{
